moved `deserialize_cheat_target` into `test_execution_syscall_handler.rs`
